requirements.txt
aiogram==3.8.0
python-dotenv==1.0.1
python_telegram_bot==21.3
psycopg2-binary==2.9.9
